Remove template-only component setting
The ClientDetail.attributesTable assertion was failing because
it becoming a template-only component meant that the
data-test-attributes selector had no element to attach to.

Having the edition set to Octane in package.json causes an error
to be thrown when template-only-glimmer-components is turned
off, so it’ll have to go for now.